requirements.txt
openai 
python-dotenv
linkup-sdk
streamlit